+++ project.sh
@@ -38,6 +38,7 @@ $VENV/bin/sumr .
 
 $PIP install doql --upgrade --quiet
 $VENV/bin/doql adopt . --format less --output app.doql.less --force
+
 $PIP install sumd --upgrade --quiet
 $VENV/bin/sumd .
 $VENV/bin/sumr .