+++ CHANGELOG.md
@@ -7,6 +7,32 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.2] - 2026-04-24
+
+### Docs
+- Update README.md
+- Update SUMD.md
+- Update SUMR.md
+- Update project/README.md
+- Update project/context.md
+
+### Test
+- Update test-gui.sh
+- Update testql-scenarios/generated-api-smoke.testql.toon.yaml
+
+### Other
+- Update .gitignore
+- Update api-gateway/src/index.js
+- Update app.doql.less
+- Update command-bus/src/index.js
+- Update generators/gen-handlebars/src/index.js
+- Update generators/gen-jinja/app.py
+- Update generators/gen-twig/index.php
+- Update postgres/init/03_content_tables.sql
+- Update project.sh
+- Update project/analysis.toon.yaml
+- ... and 19 more files
+
 ## [0.0.1] - 2026-04-24
 
 ### Docs